added one more service worker to cache networks response
